Combined/requirements.txt
--only-binary=:all:
--prefer-binary

absl-py==2.4.0
attrs==25.4.0
cffi==2.0.0
contourpy==1.3.3
cycler==0.12.1
diskcache==5.6.3
flatbuffers==20181003210633
fonttools==4.61.1
Jinja2==3.1.6
jax==0.7.1
jaxlib==0.7.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.14
moonshine-voice==0.0.49
numpy<2
opencv-python
opt_einsum==3.4.0
packaging==26.0
pillow==12.1.1
piper-tts==1.4.1
protobuf==4.25.8
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pyttsx3==2.99
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.5
tflite-runtime==2.14.0
typing_extensions==4.15.0